feat: growthPredictionFunnel 컴포넌트 추가 및 GrowthPrediction 페이지 리팩토링
